fw_att_control split into class and header
